// src/typings/avatar.ts
export type AvatarSizeType = 'xs' | 'sm' | 'lg' | 'xl' | 'xxl' | number

export enum AvatarStyleTypeEnum {
    solid = 'solid',
    outline = 'outline',
    soft = 'soft',
    simple = 'simple'
}

export enum AvatarBorderTypeEnum {
    circle = 'circle',
    square = 'square'
}

export enum AvatarVariantEnum {
    image = 'image',
    text = 'text',
    icon = 'icon'
}